requirements.txt
pandas>=1.5.0
requests>=2.28.0
together>=0.2.0
pyyaml>=6.0
python-dotenv>=0.19.0